recipe-agent-app/requirements.txt
google-adk
uvicorn
fastapi
pydantic